+++ apps/mobile/web/html-renderer/src/utilts.ts
@@ -1,0 +1,4 @@
+declare const __RN__: any
+
+// eslint-disable-next-line unicorn/no-typeof-undefined
+export const isInRn = typeof __RN__ !== "undefined"
