+++ apps/mobile/web/html-renderer/tailwind.config.ts
@@ -9,5 +9,6 @@ const rootDir = path.resolve(__dirname, "../../../..")
 export default resolveConfig({
   ...baseTwConfig,
   darkMode: "media",
+  future: { hoverOnlyWhenSupported: true },
   content: ["./src/**/*.{ts,tsx}", path.resolve(rootDir, "packages/components/src/**/*.{ts,tsx}")],
 })
