+++ apps/renderer/package.json
@@ -65,7 +65,6 @@
     "idb-keyval": "6.2.1",
     "immer": "10.1.1",
     "jotai": "2.11.3",
-    "katex": "0.16.21",
     "lethargy": "1.0.9",
     "masonic": "4.0.1",
     "mdast-util-gfm-table": "^2.0.0",
@@ -105,7 +104,6 @@
     "@follow/models": "workspace:*",
     "@follow/types": "workspace:*",
     "@follow/utils": "workspace:*",
-    "@types/katex": "0.16.7",
     "@types/node": "^22.13.0",
     "@types/react": "^19.0.8",
     "@types/react-dom": "^19.0.3",
