+++ apps/renderer/src/lib/parse-html.ts
@@ -1,5 +1,6 @@
 import { MemoedDangerousHTMLStyle } from "@follow/components/common/MemoedDangerousHTMLStyle.jsx"
 import { Checkbox } from "@follow/components/ui/checkbox/index.jsx"
+import { LazyKateX } from "@follow/components/ui/katex/lazy.js"
 import { parseHtml as parseHtmlGeneral } from "@follow/components/ui/markdown/parse-html.js"
 import type { Element } from "hast"
 import type { Components } from "hast-util-to-jsx-runtime"
@@ -8,7 +9,6 @@ import { renderToString } from "react-dom/server"
 
 import { ShadowDOM } from "~/components/common/ShadowDOM"
 import { ShikiHighLighter } from "~/components/ui/code-highlighter"
-import { LazyKateX } from "~/components/ui/katex/lazy"
 import { MarkdownBlockImage, MarkdownLink, MarkdownP } from "~/components/ui/markdown/renderers"
 import { useIsInParagraphContext } from "~/components/ui/markdown/renderers/ctx"
 import { createHeadingRenderer } from "~/components/ui/markdown/renderers/Heading"
