+++ packages/components/src/ui/katex/index.tsx
