+++ packages/components/src/ui/katex/lazy.tsx
